requirements.txt
easyocr==1.7.0
matplotlib==3.7.1
numpy==1.24.3
opencv_python_headless==4.8.0.74
Pillow==9.4.0
Pillow==10.0.0
streamlit==1.24.1
